requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0
faiss-cpu
unstructured
pillow
python-docx
python-pptx
openpyxl
streamlit
tiktoken
python-dotenv
requests
